BottomNavBar added
