+++ lib/cupertinowidgets.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class CupertinoWidgets extends StatelessWidget {
+  const CupertinoWidgets({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        color: Colors.purple,
+    );
+  }
+}
